[Gepardec/mega#264] Remove leftover public modifiers, rename package and fix formatting
