+++ src/app/modules/monthly-report/components/employee-check-confirm-comment-dialog/model/EmployeeCheckConfirmDialogData.ts
