+++ src/app/modules/office-management/office-management.module.ts
@@ -11,8 +11,12 @@ import {ProjectOverviewCardComponent} from './components/project-overview-card/p
 import {EnterpriseCardComponent} from './components/enterprise-card/enterprise-card.component';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {NgxSkeletonLoaderModule} from 'ngx-skeleton-loader';
-import { ProjectsWithoutLeadsCardComponent } from './components/projects-without-leads-card/projects-without-leads-card.component';
-import { ProjektNameWithZepLinkComponent } from './components/projekt-name-with-zep-link/projekt-name-with-zep-link.component';
+import {
+  ProjectsWithoutLeadsCardComponent
+} from './components/projects-without-leads-card/projects-without-leads-card.component';
+import {
+  ProjektNameWithZepLinkComponent
+} from './components/projekt-name-with-zep-link/projekt-name-with-zep-link.component';
 import {MatBadgeModule} from '@angular/material/badge';
 
 @NgModule({
@@ -20,17 +24,17 @@ import {MatBadgeModule} from '@angular/material/badge';
   exports: [
     EmployeeCardComponent // TODO: check if export of the container (office-management) or only of the card component is necessary
   ],
-    imports: [
-        CommonModule,
-        BrowserAnimationsModule,
-        AngularMaterialModule,
-        TranslateModule.forRoot(),
-        SharedModule,
-        FlexLayoutModule,
-        FormsModule,
-        NgxSkeletonLoaderModule,
-        MatBadgeModule
-    ]
+  imports: [
+    CommonModule,
+    BrowserAnimationsModule,
+    AngularMaterialModule,
+    TranslateModule.forRoot(),
+    SharedModule,
+    FlexLayoutModule,
+    FormsModule,
+    NgxSkeletonLoaderModule,
+    MatBadgeModule
+  ]
 })
 export class OfficeManagementModule {
 }
