+++ src/styles/root.scss
@@ -131,9 +131,11 @@ app-project-overview-card, app-project-management {
 .mt-10 {
   margin-top: 10px;
 }
+
 .mb-10 {
   margin-bottom: 10px;
 }
+
 .mr-10 {
   margin-right: 10px;
 }
